Fix hypervisor without any guest
As by https://github.com/RedHatSatellite/katello-attach-subscription/issues/68, if a host does not have any guest, and facts is not populated, fail.
The resolution is to make use of virtwho option, which will gather the facts from virtwho directly.
Still no fatal error will be given anymore after the patch, but a WARNING in stderr.
This is due to a candlepin bug which does not expose (yet) the facts for a hypervisor.